test: seperated tests & test for third parameter
